rename package print to pretty
