+++ runtime/errors.go
@@ -25,7 +25,7 @@ import (
 	"github.com/onflow/cadence/runtime/ast"
 	"github.com/onflow/cadence/runtime/common"
 	"github.com/onflow/cadence/runtime/interpreter"
-	print2 "github.com/onflow/cadence/runtime/print"
+	"github.com/onflow/cadence/runtime/pretty"
 	"github.com/onflow/cadence/runtime/sema"
 )
 
@@ -45,7 +45,7 @@ func (e Error) Unwrap() error {
 func (e Error) Error() string {
 	var sb strings.Builder
 	sb.WriteString("Execution failed:\n")
-	printErr := print2.NewErrorPrettyPrinter(&sb, false).
+	printErr := pretty.NewErrorPrettyPrinter(&sb, false).
 		// TODO: capture codes in error and include in codes argument
 		PrettyPrintError(e, "", map[string]string{})
 	if printErr != nil {
