+++ runtime/sema/errors.go
@@ -26,7 +26,7 @@ import (
 	"github.com/onflow/cadence/runtime/ast"
 	"github.com/onflow/cadence/runtime/common"
 	"github.com/onflow/cadence/runtime/errors"
-	print2 "github.com/onflow/cadence/runtime/print"
+	"github.com/onflow/cadence/runtime/pretty"
 )
 
 // astTypeConversionError
@@ -85,7 +85,7 @@ type CheckerError struct {
 func (e CheckerError) Error() string {
 	var sb strings.Builder
 	sb.WriteString("Checking failed:\n")
-	printErr := print2.NewErrorPrettyPrinter(&sb, false).
+	printErr := pretty.NewErrorPrettyPrinter(&sb, false).
 		// TODO: capture codes in error and include in codes argument
 		PrettyPrintError(e, "", map[string]string{})
 	if printErr != nil {
